+++ CHANGELOG.md
@@ -17,6 +17,10 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - Added CMForestDiscovery resource
 - Added ConvertTo-CimCMScheduleString to the ResourceHelper
 - Added CMClientStatusSettings Resource
+- Added CMBoundaryGroups Resource
+- Added ConvertTo-CimBoundaries to the ResourceHelper
+- Added Convert-BoundariesIPSubnets to the ResourceHelper
+- Added Get-BoundaryInfo to the ResourceHelper
 - Added CMAssetIntelligencePoint Resource
 
 ### Changed